[Python][Schema Update] Refactor & Update Test
- Updated tests per refactor
- Removed config.get_model_settings(). Now you have to use modelParser.get_model_settings()
