+++ python/tests/test_library_helpers.py
@@ -27,72 +27,6 @@ def ai_config_runtime():
     return runtime
 
 
-def test_get_model_settings(ai_config_runtime: AIConfigRuntime):
-    """
-    Test the get_model_settings_for_prompt method of the AIConfig class.
-    3 cases:
-    1. settings is defined as an empty dictionary
-    2. settings is defined under prompt metadata
-    3. settings is defined under config metadata. This is essentially the final default
-    4. settings is defined under config metadata, but model and not settings is defined under prompt metadata. This should default to the config metadata
-    """
-    prompt = Prompt(
-        name="Prompt1",
-        input="input doesn't matter here",
-        metadata=PromptMetadata(
-            model="fake model",
-        ),
-    )
-    ai_config_runtime.add_prompt(prompt.name, prompt)
-
-    prompt = ai_config_runtime.prompts[0]
-
-    assert ai_config_runtime.get_model_settings(prompt) == {}
-
-    # settings is defined as {}. Should be returned as {}
-    aiconfig = AIConfigRuntime(
-        name="test",
-        metadata=ConfigMetadata(**{"models": {"fakemodel": {"fake_setting": "True"}}}),
-        #  here is settings = None. This implies that settings were not passed in. Should default to global params
-        prompts=[
-            Prompt(
-                name="test",
-                input="test",
-                metadata=PromptMetadata(model=ModelMetadata(name="test", settings={})),
-            )
-        ],
-    )
-
-    prompt = aiconfig.prompts[0]
-
-    assert aiconfig.get_model_settings(prompt) == {}
-    # settings is defined as None. Should be returned as config level, ie {"fake_setting": "True"}
-    aiconfig = AIConfigRuntime(
-        name="test",
-        metadata=ConfigMetadata(**{"models": {"fakemodel": {"fake_setting": "True"}}}),
-        #  here is settings = None. This implies that settings were not passed in. Should default to global params
-        prompts=[
-            Prompt(
-                name="test",
-                input="test",
-                metadata=PromptMetadata(model=ModelMetadata(name="fakemodel", settings=None)),
-            )
-        ],
-    )
-
-    prompt = aiconfig.prompts[0]
-
-    assert aiconfig.get_model_settings(prompt) == {"fake_setting": "True"}
-
-    with pytest.raises(IndexError, match=r"Prompt '.*' not in config"):
-        prompt = Prompt(
-            name="doesn't exist",
-            input="doesn't exist",
-            metadata=PromptMetadata(model="doesn't exist"),
-        )
-        settings = aiconfig.get_model_settings(prompt)
-
-
 def test_collect_prompt_references():
     """
     Test the collection of prompt references within an AI configuration.
